// src/shared/types.ts
export interface NutrientsPer100g {
  kcal: number | null;
  protein: number | null;
  carbs: number | null;
  sugar: number | null;
  fiber: number | null;
  fat: number | null;
  satFat: number | null;
  salt: number | null;
  sodium: number | null;
}